Adds thermometer and voltmeter
thermometer is designed to be attached to amplified thermocouples, the calibrate method will eventually be made to automatically set the intercepts and slopes for the values.

voltmeter is a high-precision analog input that combines multiple analog pins for increased precision.  it also needs calibration based on the scaling factor for the fine input, as well as the voltages that the output steps at.